Handle an array of name possibilities
This makes it easier to cover different names people might put in the
signup spreadsheet without having to create duplicate entries in the
family emails list.

Also, send a notification to the admin (me!) if there's a family signed
up for something but there's no email match to send a notification to.